Fix decimal/double mismatches in calculation and tender services

+++ Persistence/Service/CalculationItems/Tender/TenderAttributeCommandService .cs
@@ -38,7 +38,7 @@ namespace Persistence.Service.CalculationItems.Tender
                         new TenderAttributeBindEntity(
                             tenderId: x.Key,
                             attributeId: attr.Id,
-                            value: x.Value));
+                            value: (decimal)x.Value));
                 }
 
                 await context.SaveChangesAsync(ct);
